+++ src/cheat_list_commands/PlaceZoneCheatListCommand.h
@@ -20,10 +20,10 @@ class PlaceZoneCheatListCommand : public ICheatListCommand
 public:
 	PlaceZoneCheatListCommand(
 		cISC4ZoneManager::ZoneType zoneType,
-		long x1,
-		long y1,
-		long x2,
-		long y2);
+		int32_t x1,
+		int32_t y1,
+		int32_t x2,
+		int32_t y2);
 
 private:
 	void Execute(
@@ -32,9 +32,9 @@ private:
 		cIGZCommandServer* pCommandServer) override;
 
 	cISC4ZoneManager::ZoneType zoneType;
-	long x1;
-	long y1;
-	long x2;
-	long y2;
+	int32_t x1;
+	int32_t y1;
+	int32_t x2;
+	int32_t y2;
 };
